requirements.txt
asgiref==3.9.0
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
Serializer==0.2.1
simplejson==3.20.1
sqlparse==0.5.3
typing_extensions==4.14.0
